+++ Demo/Demo/Keyboard/KeyboardView.swift
@@ -11,12 +11,14 @@ import SwiftUI
 
 /**
  This is the main view that is registered when the extension
- calls `setup(with:)` in `KeyboardViewController`.
+ runs `setup(with:)` in ``KeyboardViewController``. The view
+ is used by all `SystemKeyboard`-based keyboards.
  
- The view must observe the KeyboardContext as an environment
- object or inject an instance and then set it to an observed
- object (commented out). Otherwise this view will not change
- when the context changes.
+ The view must observe a `KeyboardContext` as an environment
+ object, or take a context instance as an init parameter and
+ set it to an observed object. Otherwise, it will not change
+ when the context changes. This is not how it should be, but
+ I have not yet figured out why this is needed.
  */
 struct KeyboardView: View {
     
@@ -38,7 +40,9 @@ struct KeyboardView: View {
 // MARK: - Private Views
 
 private extension KeyboardView {
-    
+
+    /// This text field can be added to the VStack above, to
+    /// test typing in a text field within the keyboard view.
     var textField: some View {
         KeyboardTextField(text: $text) {
             $0.placeholder = "Try typing here, press return to stop."