+++ Demo/Demo/Keyboard/KeyboardViewController.swift
@@ -11,47 +11,48 @@ import SwiftUI
 
 /**
  This keyboard demonstrates how to create a keyboard that is
- using `SystemKeyboard` to mimic a native iOS keyboard.
+ using `SystemKeyboard` to mimic a native English keyboard.
  
- The demo makes demo-specific configurations in `viewDidLoad`
- and registers a custom view in `viewWillSetupKeyboard`.
+ The keyboard makes demo-specific configurations and sets up
+ the keyboard with a ``KeyboardView``. You can change all of
+ these configurations to see how the keyboard changes.
  
  To use this keyboard, you must enable it in system settings
  ("Settings/General/Keyboards"). It needs full access to get
  access to features like haptic and audio feedback.
  
  Note that this demo adds KeyboardKit as a local package and
- not a remote package, as you would normally do. It makes it
- possible to change things in the library directly from this
- project instead of having to push changes to GitHub.
+ not as a remote package, as you would normally add it. This
+ is done to make it possible to change the package from this
+ project and make it easier to quickly try out new things.
  */
 class KeyboardViewController: KeyboardInputViewController {
     
     /**
-     In this demo, we will configure KeyboardKit to use some
-     demo-specific services.
+     Here, we register demo-specific services which are then
+     used by the keyboard.
      */
     override func viewDidLoad() {
         
-        // Setup a fake, demo-specific autocomplete provider
+        // Setup a fake, demo-specific autocomplete provider.
+        // 💡 You can change this provider to see how the autocomplete changes.
         autocompleteProvider = FakeAutocompleteProvider()
 
-        // Setup an alternate input set.
-        // 💡 We can use a regular input set provider with an alternate input set.
-        // inputSetProvider = EnglishInputSetProvider(
-        //     alphabetic: .englishAzerty)
+        // Setup an custom input set provider.
+        // 💡 Have a look at the other demo projects, where this is done.
+        // inputSetProvider = ...
         
-        // Setup a demo-specific keyboard appearance
-        // 💡 Play with this to change style of the keyboard
+        // Setup a demo-specific keyboard appearance.
+        // 💡 You can change this appearance to see how the keyboard style changes.
         keyboardAppearance = DemoKeyboardAppearance(context: keyboardContext)
         
-        // Setup a demo-specific keyboard action handler
-        // 💡 Play with this to change the keyboard behavior
+        // Setup a demo-specific keyboard action handler.
+        // 💡 You can change this handler to see how the keyboard behavior changes.
         keyboardActionHandler = DemoKeyboardActionHandler(
             inputViewController: self)
         
-        // Setup a demo-specific keyboard layout provider
-        // 💡 Play with this to change the keyboard's layout
+        // Setup a demo-specific keyboard layout provider.
+        // 💡 You can change this provider to see how the keyboard layout changes.
         keyboardLayoutProvider = DemoKeyboardLayoutProvider(
             inputSetProvider: inputSetProvider,
             dictationReplacement: nil)
@@ -64,13 +65,14 @@ class KeyboardViewController: KeyboardInputViewController {
      This function is called whenever the keyboard should be
      created or updated.
      
-     Here, we call setup with a demo-specific view that uses
-     ``SystemKeyboard`` to mimic a native iOS keyboard.     
+     Here, we use the ``KeyboardView`` to setup the keyboard.
+     This will create a `SystemKeyboard`-based keyboard that
+     looks like a native keyboard.
      */
     override func viewWillSetupKeyboard() {
         super.viewWillSetupKeyboard()
         
-        // Setup a demo-specific keyboard view
+        // Setup the demo with demo-specific keyboard view.
         setup(with: KeyboardView())
     }
 }